+++ examples/keycloak-e2e/server.php
@@ -112,7 +112,6 @@ $endpoint = new SecuredHttpEndpoint(
     responseFactory: $psr17,
     streamFactory: $psr17,
     allowedHosts: [ADDRESS, 'localhost:8973'],
-    maxBodyBytes: 1_048_576,
     logger: $logger,
     authentication: new BearerAuthenticationMiddleware(
         new JwksAccessTokenValidator($keys, ISSUER, RESOURCE),